+++ utilities/blob_db/blob_db.h
@@ -168,10 +168,10 @@ class BlobDB : public StackableDB {
   }
   virtual void MultiGet(const ReadOptions& /*options*/,
                         ColumnFamilyHandle* /*column_family*/,
-                        const int num_keys, const Slice* /*keys*/,
+                        const size_t num_keys, const Slice* /*keys*/,
                         PinnableSlice* /*values*/, Status* statuses,
                         const bool /*sorted_input*/ = false) override {
-    for (int i = 0; i < num_keys; ++i) {
+    for (size_t i = 0; i < num_keys; ++i) {
       statuses[i] = Status::NotSupported(
           "Blob DB doesn't support batched MultiGet");
     }
